Switch from float to integer IMU calculations
Switched the floating point IMU calculations to integers for speed and higher
resolution of orientation.

+++ Pwm.h
@@ -30,7 +30,7 @@ class Pwm {
 	void disable();
 	void set_period(int ns);
 	void set_duty_cycle(int percent);
-	void set_12dc_percent(float percent);
+	void set_12dc_percent(int percent);
 	bool ok();
 	
 	private:
